Created Entity_Object
for testing spawning objects (experimenting for now)

+++ SampleGame/SampleGame/Student.h
@@ -18,7 +18,7 @@ public:
 
 	}
 
-	int getAvg() const
+	float getAvg() const
 	{
 		return m_avg;
 	}
